Limit supported configurations to node >= 2.4.0.
We use C++11 features, which versions of node before 1.8.4 don't seem
to support (at least not with their default compiler options).

We also use `v8::NativeWeakMap`, which was introduced in the 2.4.0 ABI.
This could probably be worked around (see the node-1.8.4 branch in
github), but it doesn't seem worth it right now.