Fix crash on OSX caused by race to deallocate uv_barrier_t.
On OSX both the JS and PHP sides halt at the barrier.  The PHP side
then races to exit the barrier and deallocate the object, *before*
the JS side manages to exit the barrier.  When the JS side wakes up,
the barrier is garbage and the process crashes.

Use the return value for the uv_barrier_t as it was intended, so the
last to leave can deallocate the barrier behind it.