Improve ref counting.
Passing arrays back and forth between PHP and JavaScript means we need
to get a handle on pass-by-value and references.  Unfortunately we can't
get compile-time-pass-by-reference to work with magic methods, so tweak
so that we always separate the zvals we get to ensure they are
pass-by-value (and don't corrupt copy-on-write aliases).  We'll need to
use an explicit wrapper class to pass arrays by reference to JS.